Use fixed names for cstar_perf component logs.
The reason why we want to have fixed names instead of the generated log
file names, is that once a cstar_perf component is restarted, the log
file with the generated name will be overwritten and lost. However,
after a restart we would like to be able to still preserve the log file.